pyproject.toml
[project]
name = "terminal-space"
version = "0.1.2-dev"
description = "Default template for PDM package"
authors = [
    {name = "Don Brown", email = "mrdon@twdata.org"},
]
dependencies = [
    "aiohttp>=3.5.4",
    "colorclass>=2.0.0",
    "networkx>=1.9.1",
    "tabulate>=0.7.5",
    "prompt-toolkit>=2.0.9",
    "terminaltexteffects>=0.10.1",
    "pillow>=10.3.0",
]
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = {text = "APLv2"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "nose>=1.3.7",
    "black>=24.4.2",
    "pyinstaller>=6.7",
    "twine>=5.1.0",
    "bumpversion>=0.6.0",
    "wheel>=0.43.0",
    "pytest>=8.2.1",
]
